refactor: split fuse module into separate files
